% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penumbra.R
\name{base_pal}
\alias{base_pal}
\alias{color_pal}
\title{Return colors from the Penumbra palette}
\usage{
base_pal(n, contrast = c("balanced", "contrast+", "contrast++"))

color_pal(n, contrast = c("balanced", "contrast+", "contrast++"))
}
\arguments{
\item{n}{Number of colors (maximum 8).}

\item{contrast}{One of `"balanced"`, `"contrast+"` or `"contrast++"`.}
}
\value{
A vector of colors in #RRGGBB format
}
\description{
Return colors from the Penumbra palette
}
\examples{
old_palette <- palette()
old_pars <- par(bg = "#888888")
palette(penumbra::base_pal(8, "balanced"))
plot(1:8, pch = 19, cex = 5, col = 1:8)

par(old_pars)
palette(penumbra::color_pal(8, "balanced"))
plot(1:8, pch = 19, cex = 5, col = 1:8)

palette(penumbra::color_pal(8, "contrast+"))
plot(1:8, pch = 19, cex = 5, col = 1:8)

palette(penumbra::color_pal(8, "contrast++"))
plot(1:8, pch = 19, cex = 5, col = 1:8)

palette(old_palette)
}
