% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penumbra-package.R
\docType{package}
\name{penumbra-package}
\alias{penumbra}
\alias{penumbra-package}
\title{Penumbra - a palette inspired by nature}
\description{
This imports Neal McKee's penumbra package, available at
<https://github.com/nealmckee/penumbra>.
[base_pal()] returns grayscale-ish base colors from dark to light.
[color_pal()] returns "rainbow"-style colors.
}
\details{
Note that these palettes are not suitable for people with colourblindness.
So, only use them for your own exploratory work, not for producing display
graphics that other people have to see.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hughjonesd/penumbra}
  \item Report bugs at \url{https://github.com/hughjonesd/penumbra/issues}
}

}
\author{
\strong{Maintainer}: David Hugh-Jones \email{davidhughjones@gmail.com} [copyright holder]

Other contributors:
\itemize{
  \item Neal McKee [data contributor, copyright holder]
}

}
\keyword{internal}
