% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penumbra.R
\name{base_pal}
\alias{base_pal}
\alias{color_pal}
\title{Return colors from the Penumbra palette(s)}
\usage{
base_pal(n, contrast = c("balanced", "contrast+", "contrast++"))

color_pal(n, contrast = c("balanced", "contrast+", "contrast++"))
}
\arguments{
\item{n}{Number of colors. Between 2 and 8 for `color_pal()`, between
2 and 9 for `base_pal()`.}

\item{contrast}{One of `"balanced"`, `"contrast+"` or `"contrast++"`.}
}
\value{
A vector of colors in #RRGGBB format
}
\description{
See https://github.com/nealmckee/penumbra/ for more details about these
palettes.
}
\details{
Note that these palettes are not suitable for people with colourblindness.
So, only use them for your own exploratory work, not for producing display
graphics that other people have to see.
}
\examples{
old_palette <- palette()
old_pars <- par(bg = "#888888")
palette(penumbra::base_pal(8, "balanced"))
plot(1:8, pch = 19, cex = 5, col = 1:8)

par(old_pars)
palette(penumbra::color_pal(8, "balanced"))
plot(1:8, pch = 19, cex = 5, col = 1:8)

palette(penumbra::color_pal(8, "contrast+"))
plot(1:8, pch = 19, cex = 5, col = 1:8)

palette(penumbra::color_pal(8, "contrast++"))
plot(1:8, pch = 19, cex = 5, col = 1:8)

palette(old_palette)
}
